fix: dashboard sidebar revamp
