+++ src/components/common/Sidebar.tsx
@@ -34,18 +34,19 @@ import {
   FiMenu,
 } from "react-icons/fi";
 import { IconType } from "react-icons";
-import { BsChevronDown, BsFillCaretDownFill } from "react-icons/bs";
-import { BiBell } from "react-icons/bi";
+import { BsFillCaretDownFill } from "react-icons/bs";
+// import {  } from "react-icons/bi";
 import { FaRegEnvelope } from "react-icons/fa";
 import { RiLogoutCircleLine } from "react-icons/ri";
 import SearchBox from "../dashboard/navigation/SearchBox";
 import UserImage from "../dashboard/navigation/UserImage";
-import { useRouter } from "next/navigation";
+// import { useRouter } from "next/navigation";
 import Share from "../modals/Share";
 import useProfile from "@/hooks/useProfile";
-import { BiLogOut } from "react-icons/bi";
+import { BiLogOut, BiBell } from "react-icons/bi";
 import PageLoader from "./PageLoader";
 import useSignOut from "@/hooks/useSignOut";
+import NavItem from "./NavItem";
 // import { useLogbook } from "@/context/LogbookProvider";
 
 interface SidebarWithHeaderProps {
@@ -62,15 +63,15 @@ interface LinkItemProps {
   btnFunc?: () => void;
 }
 
-interface NavItemProps extends FlexProps {
-  icon: React.ElementType;
-  children?: React.ReactNode;
-  passedActive: string;
-  subNav: LinkItemProps[] | null;
-  navName: string;
-  href: string;
-  navType: string;
-}
+// interface NavItemProps extends FlexProps {
+//     icon: React.ElementType;
+//     children?: React.ReactNode;
+//     passedActive: string;
+//     subNav: LinkItemProps[] | null;
+//     navName: string;
+//     href: string;
+//     navType: string;
+// }
 
 interface MobileProps extends FlexProps {
   onOpen: () => void;
@@ -81,7 +82,7 @@ interface SidebarProps extends BoxProps {
   passedActive: string;
 }
 
-type RouteChangeHandler = (newRoute: string) => void;
+// type RouteChangeHandler = (newRoute: string) => void;
 
 const LinkItems: Array<LinkItemProps> = [
   { id: 1, name: "Dashboard", href: "/dashboard", icon: FiHome },
@@ -101,19 +102,19 @@ const LinkItems: Array<LinkItemProps> = [
       {
         id: 1,
         name: "Morbidity/Mortality",
-        href: "#",
+        href: "/dashboard/quality_improvement/mobility",
         icon: FiStar,
       },
       {
         id: 2,
         name: "Clinical audit",
-        href: "#",
+        href: "/dashboard/quality_improvement/clinical_audit",
         icon: FiStar,
       },
       {
         id: 3,
         name: "Case review",
-        href: "#",
+        href: "/dashboard/quality_improvement/case_review",
         icon: FiStar,
       },
     ],
@@ -203,7 +204,7 @@ const SidebarContent = ({ onClose, passedActive, ...rest }: SidebarProps) => {
         {LinkItems.map((link) => (
           <NavItem
             key={link.name}
-            navType={link.name}
+            // navType={link.name}
             icon={link.icon}
             subNav={link.children ?? null}
             navName={link.name}
@@ -251,124 +252,141 @@ const SidebarContent = ({ onClose, passedActive, ...rest }: SidebarProps) => {
   );
 };
 
-const NavItem = ({
-  icon,
-  passedActive,
-  subNav,
-  navName,
-  href,
-  navType,
-  ...rest
-}: NavItemProps) => {
-  const router = useRouter();
-  // const { handleLogbookMode } = useLogbook();
-  const [showSubNav, setShowSubNav] = useState<boolean>(false);
-  const handleRouteChange: RouteChangeHandler = (newRoute) => {
-    router.push(newRoute);
-  };
+// const NavItem = ({
+//     icon,
+//     passedActive,
+//     subNav,
+//     navName,
+//     href,
+//     navType,
+//     ...rest
+// }: NavItemProps) => {
+//     const router = useRouter();
+//     // const { handleLogbookMode } = useLogbook();
+//     const [showSubNav, setShowSubNav] = useState<boolean>(false);
+//     const handleRouteChange: RouteChangeHandler = (newRoute) => {
+//         router.push(newRoute);
+//     };
 
-  const handleShowSubNav = () => {
-    setShowSubNav(!showSubNav);
-  };
+//     const handleShowSubNav = () => {
+//         setShowSubNav(!showSubNav);
+//     };
 
-  return (
-    <Box>
-      <Box
-        onClick={() => {
-          if (subNav) {
-            handleShowSubNav();
-          } else {
-            handleRouteChange(href);
-          }
-        }}
-        style={{ textDecoration: "none" }}
-        _focus={{ boxShadow: "none" }}
-      >
-        <Flex
-          align="center"
-          p="4"
-          px="4"
-          borderRadius="lg"
-          role="group"
-          cursor="pointer"
-          color="grey_1"
-          bgColor={passedActive === href || showSubNav ? "white" : "grey_9"}
-          _hover={{
-            bg: "white",
-          }}
-          gap="0.9rem"
-          {...rest}
-        >
-          {icon && (
-            <Flex
-              align="center"
-              justify="center"
-              minW="2rem"
-              minH="2rem"
-              rounded="full"
-              bgColor={
-                passedActive === href || showSubNav ? "primary" : "white"
-              }
-              _groupHover={{
-                bgColor: "primary",
-              }}
-            >
-              <Icon
-                fontSize="16"
-                _groupHover={{
-                  color: "white",
-                }}
-                as={icon}
-                color={
-                  passedActive === href || showSubNav ? "white" : "primary"
-                }
-              />
-            </Flex>
-          )}
-          <Text
-            fontSize="0.9375rem"
-            color={passedActive === href || showSubNav ? "black" : "grey_1"}
-            fontWeight={passedActive === href || showSubNav ? "700" : "normal"}
-          >
-            {navName}
-          </Text>
+//     return (
+//         <Box>
+//             <Box
+//                 onClick={() => {
+//                     if (subNav) {
+//                         handleShowSubNav();
+//                     } else {
+//                         handleRouteChange(href);
+//                     }
+//                 }}
+//                 style={{ textDecoration: "none" }}
+//                 _focus={{ boxShadow: "none" }}
+//             >
+//                 <Flex
+//                     align="center"
+//                     p="4"
+//                     px="4"
+//                     borderRadius="lg"
+//                     role="group"
+//                     cursor="pointer"
+//                     color="grey_1"
+//                     bgColor={
+//                         passedActive === href || showSubNav ? "white" : "grey_9"
+//                     }
+//                     _hover={{
+//                         bg: "white",
+//                     }}
+//                     gap="0.9rem"
+//                     {...rest}
+//                 >
+//                     {icon && (
+//                         <Flex
+//                             align="center"
+//                             justify="center"
+//                             minW="2rem"
+//                             minH="2rem"
+//                             rounded="full"
+//                             bgColor={
+//                                 passedActive === href || showSubNav
+//                                     ? "primary"
+//                                     : "white"
+//                             }
+//                             _groupHover={{
+//                                 bgColor: "primary",
+//                             }}
+//                         >
+//                             <Icon
+//                                 fontSize="16"
+//                                 _groupHover={{
+//                                     color: "white",
+//                                 }}
+//                                 as={icon}
+//                                 color={
+//                                     passedActive === href || showSubNav
+//                                         ? "white"
+//                                         : "primary"
+//                                 }
+//                             />
+//                         </Flex>
+//                     )}
+//                     <Text
+//                         fontSize="0.9375rem"
+//                         color={
+//                             passedActive === href || showSubNav
+//                                 ? "black"
+//                                 : "grey_1"
+//                         }
+//                         fontWeight={
+//                             passedActive === href || showSubNav
+//                                 ? "700"
+//                                 : "normal"
+//                         }
+//                     >
+//                         {navName}
+//                     </Text>
 
-          {subNav && <Icon as={BsChevronDown} />}
-        </Flex>
-      </Box>
-      {subNav && (
-        <>
-          {showSubNav && (
-            <Box>
-              {subNav.map((item) => (
-                <Box
-                  pl="4rem"
-                  _hover={{
-                    bgColor: "white",
-                    cursor: "pointer",
-                  }}
-                  key={item?.name}
-                  fontSize="0.9375rem"
-                  py="0.47rem"
-                  mb="0.47rem"
-                  onClick={() => {
-                    if (navType === "Logbook" && item.href) {
-                      // handleLogbookMode(item?.name);
-                      handleRouteChange(item.href);
-                    }
-                  }}
-                >
-                  {" "}
-                  {item?.name}{" "}
-                </Box>
-              ))}
-            </Box>
-          )}
-        </>
-      )}
-    </Box>
-  );
-};
+//                     {subNav && <Icon as={BsChevronDown} />}
+//                 </Flex>
+//             </Box>
+//             {subNav && (
+//                 <>
+//                     {showSubNav && (
+//                         <Box>
+//                             {subNav.map((item) => (
+//                                 <Box
+//                                     pl="4rem"
+//                                     _hover={{
+//                                         bgColor: "white",
+//                                         cursor: "pointer",
+//                                     }}
+//                                     key={item?.name}
+//                                     fontSize="0.9375rem"
+//                                     py="0.47rem"
+//                                     mb="0.47rem"
+//                                     onClick={() => {
+//                                         if (
+//                                             navType === "Logbook" &&
+//                                             item.href
+//                                         ) {
+//                                             // handleLogbookMode(item?.name);
+//                                             handleRouteChange(item.href);
+//                                         }
+//                                     }}
+//                                 >
+//                                     {" "}
+//                                     {item?.name}{" "}
+//                                 </Box>
+//                             ))}
+//                         </Box>
+//                     )}
+//                 </>
+//             )}
+//         </Box>
+//     );
+// };
 
 const MobileNav = ({ onOpen, ...rest }: MobileProps) => {
   const { data } = useProfile();
